// src/pages/Settings/Objects/types.ts
export type ObjectItem = {
    id: number;
    name: string;
    company_id: number;
    region_id: number;
    district_id: number;
    address: string;
    start_date: string;
    end_date: string;
    floor_count: number;
    block_count: number;
    status: number;
    is_active: number;
    created_at: string;
    updated_at: string;
};